Introduce rec-main.hh and split out TCP code to rec-tcp.cc
