+++ pdns/recursordist/Makefile.am
@@ -160,10 +160,12 @@ pdns_recursor_SOURCES = \
 	rcpgenerator.cc rcpgenerator.hh \
 	rec-carbon.cc \
 	rec-eventtrace.cc rec-eventtrace.hh \
+	ec-main.hh rec-main.cc \
 	rec-lua-conf.hh rec-lua-conf.cc \
 	rec-protozero.cc rec-protozero.hh \
 	rec-snmp.hh rec-snmp.cc \
 	rec-taskqueue.cc rec-taskqueue.hh \
+	rec-tcp.cc \
         rec-tcpout.cc rec-tcpout.hh \
         rec-zonetocache.cc rec-zonetocache.hh \
 	rec_channel.cc rec_channel.hh rec_metrics.hh \
